backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
pydantic-settings==2.10.1
PyJWT==2.10.1
cryptography==46.0.6
